# vectorize 'auto' long decoding