--- a/processing/src/main/java/org/apache/druid/segment/data/DeltaLongEncodingReader.java
+++ b/processing/src/main/java/org/apache/druid/segment/data/DeltaLongEncodingReader.java
@@ -68,23 +68,13 @@ public class DeltaLongEncodingReader implements CompressionFactory.LongEncodingR
   @Override
   public void read(long[] out, int outPosition, int startIndex, int length)
   {
-    deserializer.get(out, outPosition, startIndex, length);
-
-    for (int i = 0; i < length; i++) {
-      out[outPosition + i] += base;
-    }
+    deserializer.getDelta(out, outPosition, startIndex, length, base);
   }
 
   @Override
   public int read(long[] out, int outPosition, int[] indexes, int length, int indexOffset, int limit)
   {
-    final int len = deserializer.get(out, outPosition, indexes, length, indexOffset, limit);
-
-    for (int i = 0; i < len; i++) {
-      out[outPosition + i] += base;
-    }
-
-    return len;
+    return deserializer.getDelta(out, outPosition, indexes, length, indexOffset, limit, base);
   }
 
   @Override