--- a/processing/src/main/java/org/apache/druid/segment/data/DeltaLongEncodingReader.java
+++ b/processing/src/main/java/org/apache/druid/segment/data/DeltaLongEncodingReader.java
@@ -66,6 +66,28 @@ public class DeltaLongEncodingReader implements CompressionFactory.LongEncodingR
   }
 
   @Override
+  public void read(long[] out, int outPosition, int startIndex, int length)
+  {
+    deserializer.get(out, outPosition, startIndex, length);
+
+    for (int i = 0; i < length; i++) {
+      out[outPosition + i] += base;
+    }
+  }
+
+  @Override
+  public int read(long[] out, int outPosition, int[] indexes, int length, int indexOffset, int limit)
+  {
+    final int len = deserializer.get(out, outPosition, indexes, length, indexOffset, limit);
+
+    for (int i = 0; i < len; i++) {
+      out[outPosition + i] += base;
+    }
+
+    return len;
+  }
+
+  @Override
   public CompressionFactory.LongEncodingReader duplicate()
   {
     return new DeltaLongEncodingReader(buffer.duplicate(), base, bitsPerValue);