--- a/processing/src/main/java/org/apache/druid/segment/data/TableLongEncodingReader.java
+++ b/processing/src/main/java/org/apache/druid/segment/data/TableLongEncodingReader.java
@@ -72,6 +72,28 @@ public class TableLongEncodingReader implements CompressionFactory.LongEncodingR
   }
 
   @Override
+  public void read(long[] out, int outPosition, int startIndex, int length)
+  {
+    deserializer.get(out, outPosition, startIndex, length);
+
+    for (int i = 0; i < length; i++) {
+      out[outPosition + i] = table[(int) out[outPosition + i]];
+    }
+  }
+
+  @Override
+  public int read(long[] out, int outPosition, int[] indexes, int length, int indexOffset, int limit)
+  {
+    final int len = deserializer.get(out, outPosition, indexes, length, indexOffset, limit);
+
+    for (int i = 0; i < len; i++) {
+      out[outPosition + i] = table[(int) out[outPosition + i]];
+    }
+
+    return len;
+  }
+
+  @Override
   public CompressionFactory.LongEncodingReader duplicate()
   {
     return new TableLongEncodingReader(buffer.duplicate(), table, bitsPerValue);