### Description
This PR specializes `LongDeserializer` implementations with `getDelta` and `getTable` methods to push down unpacking bits, delta encoding adjustment, and table lookups for vectors of data as far as possible and work more efficiently with vectorized query engines, primarily focusing on contiguous reads.

It works by unrolling value reads to line up with `ByteBuffer` get methods to eliminate overlapping reads where possible, reading blocks of 8 values at a time for un-aligned values (1, 2, 4, 12, 20, 24, 40, 48, 56). `ByteBuffer` aligned bit-packing widths (8, 16, 32, 64) actually performed worse when this same unrolling was performed, so instead they utilize a traditional for loop with the aligned get methods.

Most of the improvement is on the small end since it has the most redundant/overlapping memory accesses, but is pretty decent improvement there.

### Full column scans before/after on uniform distribution columns of varying bits to cover the entire set of value decoders
before:

```
Benchmark                                                                           (distribution)  (encoding)  (filteredRowCountPercentage)   (rows)  (zeroProbability)  Mode  Cnt         Score      Error  Units
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                       uniform-1    lz4-auto                           1.0  5000000                0.0  avgt    5     23862.994 ± 1944.514  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                       uniform-2    lz4-auto                           1.0  5000000                0.0  avgt    5     23567.778 ±  757.632  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                       uniform-4    lz4-auto                           1.0  5000000                0.0  avgt    5     28826.685 ± 3260.379  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                       uniform-8    lz4-auto                           1.0  5000000                0.0  avgt    5     19829.688 ±  923.047  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                      uniform-12    lz4-auto                           1.0  5000000                0.0  avgt    5     32419.941 ± 1486.951  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                      uniform-16    lz4-auto                           1.0  5000000                0.0  avgt    5     22460.206 ± 4207.942  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                      uniform-20    lz4-auto                           1.0  5000000                0.0  avgt    5     31415.040 ± 3867.699  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                      uniform-24    lz4-auto                           1.0  5000000                0.0  avgt    5     24444.519 ± 2472.537  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                      uinform-32    lz4-auto                           1.0  5000000                0.0  avgt    5     18031.908 ± 1471.220  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                      uniform-40    lz4-auto                           1.0  5000000                0.0  avgt    5     22442.611 ± 1778.178  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                      uniform-48    lz4-auto                           1.0  5000000                0.0  avgt    5     24733.015 ± 3076.896  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                      uniform-56    lz4-auto                           1.0  5000000                0.0  avgt    5     23201.264 ± 1578.009  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                      uniform-64    lz4-auto                           1.0  5000000                0.0  avgt    5     20581.399 ± 1269.190  us/op
```

after:
```
Benchmark                                                                           (distribution)  (encoding)  (filteredRowCountPercentage)   (rows)
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                       uniform-1    lz4-auto                           1.0  5000000                0.0  avgt    5     17892.362 ±  102.776  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                       uniform-2    lz4-auto                           1.0  5000000                0.0  avgt    5     17796.103 ±  417.847  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                       uniform-4    lz4-auto                           1.0  5000000                0.0  avgt    5     17879.496 ±  237.066  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                       uniform-8    lz4-auto                           1.0  5000000                0.0  avgt    5     17508.260 ±  560.856  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                      uniform-12    lz4-auto                           1.0  5000000                0.0  avgt    5     18272.440 ±   71.751  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                      uniform-16    lz4-auto                           1.0  5000000                0.0  avgt    5     19042.292 ±  595.685  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                      uniform-20    lz4-auto                           1.0  5000000                0.0  avgt    5     18782.746 ±  248.738  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                      uniform-24    lz4-auto                           1.0  5000000                0.0  avgt    5     19048.354 ±  160.025  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                      uinform-32    lz4-auto                           1.0  5000000                0.0  avgt    5     17984.778 ±  633.691  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                      uniform-40    lz4-auto                           1.0  5000000                0.0  avgt    5     22070.007 ±  166.035  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                      uniform-48    lz4-auto                           1.0  5000000                0.0  avgt    5     22052.517 ± 2168.763  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                      uniform-56    lz4-auto                           1.0  5000000                0.0  avgt    5     25001.739 ±  259.184  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                      uniform-64    lz4-auto                           1.0  5000000                0.0  avgt    5     20325.369 ±   60.724  us/op
```

### Full column scans on other value distribution before/after comparison
before:
```
Benchmark                                                                           (distribution)  (encoding)  (filteredRowCountPercentage)   (rows)  (zeroProbability)  Mode  Cnt         Score      Error  Units
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                  enumerated-0-1    lz4-auto                           1.0  5000000                0.0  avgt    5     24186.788 ± 1440.692  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                 enumerated-full    lz4-auto                           1.0  5000000                0.0  avgt    5     27220.865 ± 2221.740  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                     normal-1-32    lz4-auto                           1.0  5000000                0.0  avgt    5     19490.250 ± 1231.429  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                  normal-40-1000    lz4-auto                           1.0  5000000                0.0  avgt    5     20701.792 ± 1898.845  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                 sequential-1000    lz4-auto                           1.0  5000000                0.0  avgt    5     28977.482 ±  878.240  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized               sequential-unique    lz4-auto                           1.0  5000000                0.0  avgt    5     22856.419 ±  247.786  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                    zipf-low-100    lz4-auto                           1.0  5000000                0.0  avgt    5     19135.244 ±  437.260  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                 zipf-low-100000    lz4-auto                           1.0  5000000                0.0  avgt    5     31696.149 ± 2933.263  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                 zipf-low-32-bit    lz4-auto                           1.0  5000000                0.0  avgt    5     25815.759 ± 2073.480  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                   zipf-high-100    lz4-auto                           1.0  5000000                0.0  avgt    5     20582.544 ± 1726.414  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                zipf-high-100000    lz4-auto                           1.0  5000000                0.0  avgt    5     18996.105 ± 1412.307  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                zipf-high-32-bit    lz4-auto                           1.0  5000000                0.0  avgt    5     18628.002 ±  450.125  us/op
```

after:
```
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                  enumerated-0-1    lz4-auto                           1.0  5000000                0.0  avgt    5     19250.882 ± 1221.067  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                 enumerated-full    lz4-auto                           1.0  5000000                0.0  avgt    5     19853.160 ±  780.701  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                     normal-1-32    lz4-auto                           1.0  5000000                0.0  avgt    5     18414.646 ± 1705.926  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                  normal-40-1000    lz4-auto                           1.0  5000000                0.0  avgt    5     20054.183 ± 2539.912  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                 sequential-1000    lz4-auto                           1.0  5000000                0.0  avgt    5     18585.123 ± 1524.581  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized               sequential-unique    lz4-auto                           1.0  5000000                0.0  avgt    5     19304.239 ±  808.152  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                    zipf-low-100    lz4-auto                           1.0  5000000                0.0  avgt    5     19248.088 ±  470.994  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                 zipf-low-100000    lz4-auto                           1.0  5000000                0.0  avgt    5     23815.318 ± 4413.581  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                 zipf-low-32-bit    lz4-auto                           1.0  5000000                0.0  avgt    5     26111.974 ± 4594.582  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                   zipf-high-100    lz4-auto                           1.0  5000000                0.0  avgt    5     20144.349 ±  489.978  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                zipf-high-100000    lz4-auto                           1.0  5000000                0.0  avgt    5     18956.415 ±  796.136  us/op
ColumnarLongsSelectRowsFromGeneratorBenchmark.selectRowsVectorized                zipf-high-32-bit    lz4-auto                           1.0  5000000                0.0  avgt    5     18666.583 ±  385.407  us/op
```

### Full benchmarks on column scan value selects with simulated filters (up to full scan) 
For the top line graph, the x axis is percent of rows selected by column scan (offset analog) and goes up to 1.0/contiguous scan. The y axis is the amount of time it took to select that many values. The numbers are kind of noisy because i didn't spend the multiple days necessary to do it proper, but it gives a decent idea. Nearly all of the improvement is on the full scan end of the spectrum (last datapoint), because the underlying `BitmapVectorOffset` will never call the contiguous vectorized get methods (which I will fix in a follow-up PR) and the contiguous gets were the primary area of improvement in this PR. The bottom bar chart is the size of the column in bytes as it would be stored in a segment (encoded and/or compressed).

![before-after](https://user-images.githubusercontent.com/1577461/111410582-3d80ce00-8696-11eb-8f6c-c57230ba0849.gif)

In a follow-up PR i will probably argue for making this the default long encoding since it does pretty well with the vectorized engine, compared to lz4 longs.

<hr>

### Unpacking
I uh, drew pictures to help me get the shifts and masking correct, so I'll add them here in case they are any help to reviewers. Imagine each diagram for the bit width is the number of values to unpack (8) per loop, where each color is a separate value, and each box is 4 bits.

<img width="1058" alt="PNG image-9F52520DEE1D-1" src="https://user-images.githubusercontent.com/1577461/111421585-7d51b080-86aa-11eb-8360-933e60274e3c.png">
<img width="1036" alt="PNG image-BA4F111C2B53-1" src="https://user-images.githubusercontent.com/1577461/111421614-86db1880-86aa-11eb-9651-cc750878125a.png">
<img width="958" alt="PNG image-F37823656142-1" src="https://user-images.githubusercontent.com/1577461/111421645-94909e00-86aa-11eb-9127-8f37c403e727.png">
<img width="974" alt="PNG image-8754DDE6B474-1" src="https://user-images.githubusercontent.com/1577461/111421668-9ce8d900-86aa-11eb-963f-79b5521ab92e.png">
<img width="1022" alt="PNG image-D07C1140D960-1" src="https://user-images.githubusercontent.com/1577461/111421682-a3775080-86aa-11eb-85d2-e84ab0e12885.png">


<hr>

This PR has:
- [x] been self-reviewed.
- [ ] added documentation for new or modified features or behaviors.
- [x] added Javadocs for most classes and all non-trivial methods. Linked related entities via Javadoc links.
- [x] added comments explaining the "why" and the intent of the code wherever would not be obvious for an unfamiliar reader.
- [x] added unit tests or modified existing tests to cover new code paths, ensuring the threshold for [code coverage](https://github.com/apache/druid/blob/master/dev/code-review/code-coverage.md) is met.
- [ ] been tested in a test Druid cluster.